Added NHLScraper module
Added a module to scrape the nhl.com REST API for data. The main.py script now gets the daily schedule, and outputs the games to the LCD. This is currently using the GoalCelebration module for this, which will probably be renamed in the next commit, but for now it seems to be working. No error handling currently, so if there are no games on the given day, you're presented with an error message.