Added NHLScoringPlays class
Created a new class called NHLScoringPlays to keep track of the team who scored, when the goal was scored, the updated score, etc. Also added a basic NHLTeam class to better handle team objects, including their abbreviations. The main program now grabs the first game on a particular day, and displays all of the goals that were scored during the game. The code is still a bit of a mess, but I'm going to get all of the major logic put into the project first, and then refactor and cleanup.