Created Initial Database Structure
Created a simple database structure to contain data about Teams, Franchises, Divisions and Conferences. This will allow me to cache the data with one call to the API instead of making repeated calls, which will greatly increase the speed (last commit saw a drastic slow down). Still need to add more tables, but this is a start.