Created class to handle Franchises
Created a class to work with Franchise data. This class can construct the object from JSON data, an ID to be passed to the database, or an ID to be passed to the API. This was written very quickly, I will be reviewing it before creating other classes like it.