Improved NHLTeam class
Brought NHLTeam inline with classes like NHLDivision, with the ability to query a local cached database of team data, or query the API directly. This involved some changes of some other files, however everything is now working much faster as it is making only one call to the API, and the rest of the data comes from the local database.