Added Button class
Added a class for a button that can be pressed to initiate a function. The first button to be implemented will be for displaying the NHL schedule for the day. Currently it does nothing except print to the console that the button has been pressed. There is some commented out code in the class, this is because I'm not certain whether the function to call will be passed to the Button constructor as a parameter, or if I'll use the Button class as in interface and extend the class for each kind of button I want to have, overriding the action method. I'm somewhat leaning to the interface option right now (though you couldn't tell from my code), but it's still up in the air. Also, somewhat unrelated, I made the SportsTicker constructor parameter lcdPinBacklight mandatory.